+++ sdk/include/opentelemetry/sdk/trace/batch_span_processor.h
@@ -105,9 +105,9 @@ public:
    * shared ownership of the processor, and thus doesn't call Shutdown (as the processor might be
    * shared with other Tracers).
    */
-  ~BatchSpanProcessor();
+  virtual ~BatchSpanProcessor();
 
-private:
+protected:
   /**
    * The background routine performed by the worker thread.
    */
@@ -116,12 +116,8 @@ private:
   /**
    * Exports all ended spans to the configured exporter.
    *
-   * @param was_force_flush_called - A flag to check if the current export is the result
-   *                                 of a call to ForceFlush method. If true, then we have to
-   *                                 notify the main thread to wake it up in the ForceFlush
-   *                                 method.
    */
-  void Export(const bool was_for_flush_called);
+  virtual void Export();
 
   /**
    * Called when Shutdown() is invoked. Completely drains the queue of all its ended spans and
@@ -129,6 +125,31 @@ private:
    */
   void DrainQueue();
 
+  struct SynchronizationData
+  {
+    /* Synchronization primitives */
+    std::condition_variable cv, force_flush_cv;
+    std::mutex cv_m, force_flush_cv_m, shutdown_m;
+
+    /* Important boolean flags to handle the workflow of the processor */
+    std::atomic<bool> is_force_wakeup_background_worker;
+    std::atomic<bool> is_force_flush_pending;
+    std::atomic<bool> is_force_flush_notified;
+    std::atomic<bool> is_shutdown;
+  };
+
+  /**
+   * @brief Notify completion of shutdown and force flush. This may be called from the any thread at
+   * any time
+   *
+   * @param notify_force_flush Flag to indicate whether to notify force flush completion.
+   * @param synchronization_data Synchronization data to be notified.
+   */
+  static void NotifyCompletion(bool notify_force_flush,
+                               const std::shared_ptr<SynchronizationData> &synchronization_data);
+
+  void GetWaitAdjustedTime(std::chrono::microseconds &timeout,
+                           std::chrono::time_point<std::chrono::system_clock> &start_time);
   /* The configured backend exporter */
   std::unique_ptr<SpanExporter> exporter_;
 
@@ -137,17 +158,10 @@ private:
   const std::chrono::milliseconds schedule_delay_millis_;
   const size_t max_export_batch_size_;
 
-  /* Synchronization primitives */
-  std::condition_variable cv_, force_flush_cv_;
-  std::mutex cv_m_, force_flush_cv_m_, shutdown_m_;
-
   /* The buffer/queue to which the ended spans are added */
   common::CircularBuffer<Recordable> buffer_;
 
-  /* Important boolean flags to handle the workflow of the processor */
-  std::atomic<bool> is_shutdown_{false};
-  std::atomic<bool> is_force_flush_{false};
-  std::atomic<bool> is_force_flush_notified_{false};
+  std::shared_ptr<SynchronizationData> synchronization_data_;
 
   /* The background worker thread */
   std::thread worker_thread_;