+++ sdk/src/trace/batch_span_processor.cc
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "opentelemetry/sdk/trace/batch_span_processor.h"
+#include "opentelemetry/common/spin_lock_mutex.h"
 
 #include <vector>
 using opentelemetry::sdk::common::AtomicUniquePtr;
@@ -14,6 +15,7 @@ namespace sdk
 {
 namespace trace
 {
+
 BatchSpanProcessor::BatchSpanProcessor(std::unique_ptr<SpanExporter> &&exporter,
                                        const BatchSpanProcessorOptions &options)
     : exporter_(std::move(exporter)),
@@ -21,8 +23,14 @@ BatchSpanProcessor::BatchSpanProcessor(std::unique_ptr<SpanExporter> &&exporter,
       schedule_delay_millis_(options.schedule_delay_millis),
       max_export_batch_size_(options.max_export_batch_size),
       buffer_(max_queue_size_),
+      synchronization_data_(std::make_shared<SynchronizationData>()),
       worker_thread_(&BatchSpanProcessor::DoBackgroundWork, this)
-{}
+{
+  synchronization_data_->is_force_wakeup_background_worker.store(false);
+  synchronization_data_->is_force_flush_pending.store(false);
+  synchronization_data_->is_force_flush_notified.store(false);
+  synchronization_data_->is_shutdown.store(false);
+}
 
 std::unique_ptr<Recordable> BatchSpanProcessor::MakeRecordable() noexcept
 {
@@ -36,7 +44,7 @@ void BatchSpanProcessor::OnStart(Recordable &, const SpanContext &) noexcept
 
 void BatchSpanProcessor::OnEnd(std::unique_ptr<Recordable> &&span) noexcept
 {
-  if (is_shutdown_.load() == true)
+  if (synchronization_data_->is_shutdown.load() == true)
   {
     return;
   }
@@ -48,39 +56,83 @@ void BatchSpanProcessor::OnEnd(std::unique_ptr<Recordable> &&span) noexcept
 
   // If the queue gets at least half full a preemptive notification is
   // sent to the worker thread to start a new export cycle.
-  if (buffer_.size() >= max_queue_size_ / 2)
+  size_t buffer_size = buffer_.size();
+  if (buffer_size >= max_queue_size_ / 2 || buffer_size >= max_export_batch_size_)
   {
     // signal the worker thread
-    cv_.notify_one();
+    synchronization_data_->cv.notify_one();
   }
 }
 
 bool BatchSpanProcessor::ForceFlush(std::chrono::microseconds timeout) noexcept
 {
-  if (is_shutdown_.load() == true)
+  if (synchronization_data_->is_shutdown.load() == true)
   {
     return false;
   }
 
-  is_force_flush_ = true;
-
-  // Keep attempting to wake up the worker thread
-  while (is_force_flush_.load() == true)
-  {
-    cv_.notify_one();
-  }
-
   // Now wait for the worker thread to signal back from the Export method
-  std::unique_lock<std::mutex> lk(force_flush_cv_m_);
-  while (is_force_flush_notified_.load() == false)
+  std::unique_lock<std::mutex> lk_cv(synchronization_data_->force_flush_cv_m);
+
+  synchronization_data_->is_force_flush_pending.store(true, std::memory_order_release);
+  auto break_condition = [this]() {
+    if (synchronization_data_->is_shutdown.load() == true)
+    {
+      return true;
+    }
+
+    // Wake up the worker thread once.
+    if (synchronization_data_->is_force_flush_pending.load(std::memory_order_acquire))
+    {
+      synchronization_data_->is_force_wakeup_background_worker.store(true,
+                                                                     std::memory_order_release);
+      synchronization_data_->cv.notify_one();
+    }
+
+    return synchronization_data_->is_force_flush_notified.load(std::memory_order_acquire);
+  };
+
+  // Fix timeout to meet requirement of wait_for
+  timeout = opentelemetry::common::DurationUtil::AdjustWaitForTimeout(
+      timeout, std::chrono::microseconds::zero());
+  bool result;
+  if (timeout <= std::chrono::microseconds::zero())
   {
-    force_flush_cv_.wait(lk);
+    bool wait_result = false;
+    while (!wait_result)
+    {
+      // When is_force_flush_notified.store(true) and force_flush_cv.notify_all() is called
+      // between is_force_flush_pending.load() and force_flush_cv.wait(). We must not wait
+      // for ever
+      wait_result = synchronization_data_->force_flush_cv.wait_for(lk_cv, schedule_delay_millis_,
+                                                                   break_condition);
+    }
+    result = true;
+  }
+  else
+  {
+    result = synchronization_data_->force_flush_cv.wait_for(lk_cv, timeout, break_condition);
   }
 
-  // Notify the worker thread
-  is_force_flush_notified_ = false;
+  // If it will be already signaled, we must wait util notified.
+  // We use a spin lock here
+  if (false ==
+      synchronization_data_->is_force_flush_pending.exchange(false, std::memory_order_acq_rel))
+  {
+    for (int retry_waiting_times = 0;
+         false == synchronization_data_->is_force_flush_notified.load(std::memory_order_acquire);
+         ++retry_waiting_times)
+    {
+      opentelemetry::common::SpinLockMutex::fast_yield();
+      if ((retry_waiting_times & 127) == 127)
+      {
+        std::this_thread::yield();
+      }
+    }
+  }
+  synchronization_data_->is_force_flush_notified.store(false, std::memory_order_release);
 
-  return true;
+  return result;
 }
 
 void BatchSpanProcessor::DoBackgroundWork()
@@ -90,39 +142,26 @@ void BatchSpanProcessor::DoBackgroundWork()
   while (true)
   {
     // Wait for `timeout` milliseconds
-    std::unique_lock<std::mutex> lk(cv_m_);
-    cv_.wait_for(lk, timeout);
+    std::unique_lock<std::mutex> lk(synchronization_data_->cv_m);
+    synchronization_data_->cv.wait_for(lk, timeout, [this] {
+      if (synchronization_data_->is_force_wakeup_background_worker.load(std::memory_order_acquire))
+      {
+        return true;
+      }
 
-    if (is_shutdown_.load() == true)
+      return !buffer_.empty();
+    });
+    synchronization_data_->is_force_wakeup_background_worker.store(false,
+                                                                   std::memory_order_release);
+
+    if (synchronization_data_->is_shutdown.load() == true)
     {
       DrainQueue();
       return;
     }
 
-    bool was_force_flush_called = is_force_flush_.load();
-
-    // Check if this export was the result of a force flush.
-    if (was_force_flush_called == true)
-    {
-      // Since this export was the result of a force flush, signal the
-      // main thread that the worker thread has been notified
-      is_force_flush_ = false;
-    }
-    else
-    {
-      // If the buffer was empty during the entire `timeout` time interval,
-      // go back to waiting. If this was a spurious wake-up, we export only if
-      // `buffer_` is not empty. This is acceptable because batching is a best
-      // mechanism effort here.
-      if (buffer_.empty() == true)
-      {
-        timeout = schedule_delay_millis_;
-        continue;
-      }
-    }
-
     auto start = std::chrono::steady_clock::now();
-    Export(was_force_flush_called);
+    Export();
     auto end      = std::chrono::steady_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start);
 
@@ -131,68 +170,112 @@ void BatchSpanProcessor::DoBackgroundWork()
   }
 }
 
-void BatchSpanProcessor::Export(const bool was_force_flush_called)
+void BatchSpanProcessor::Export()
 {
-  std::vector<std::unique_ptr<Recordable>> spans_arr;
-
-  size_t num_spans_to_export;
-
-  if (was_force_flush_called == true)
+  do
   {
-    num_spans_to_export = buffer_.size();
-  }
-  else
-  {
-    num_spans_to_export =
-        buffer_.size() >= max_export_batch_size_ ? max_export_batch_size_ : buffer_.size();
-  }
-
-  buffer_.Consume(num_spans_to_export,
-                  [&](CircularBufferRange<AtomicUniquePtr<Recordable>> range) noexcept {
-                    range.ForEach([&](AtomicUniquePtr<Recordable> &ptr) {
-                      std::unique_ptr<Recordable> swap_ptr = std::unique_ptr<Recordable>(nullptr);
-                      ptr.Swap(swap_ptr);
-                      spans_arr.push_back(std::unique_ptr<Recordable>(swap_ptr.release()));
-                      return true;
-                    });
-                  });
-
-  exporter_->Export(nostd::span<std::unique_ptr<Recordable>>(spans_arr.data(), spans_arr.size()));
-
-  // Notify the main thread in case this export was the result of a force flush.
-  if (was_force_flush_called == true)
-  {
-    is_force_flush_notified_ = true;
-    while (is_force_flush_notified_.load() == true)
+    std::vector<std::unique_ptr<Recordable>> spans_arr;
+    size_t num_records_to_export;
+    bool notify_force_flush =
+        synchronization_data_->is_force_flush_pending.exchange(false, std::memory_order_acq_rel);
+    if (notify_force_flush)
     {
-      force_flush_cv_.notify_one();
+      num_records_to_export = buffer_.size();
     }
+    else
+    {
+      num_records_to_export =
+          buffer_.size() >= max_export_batch_size_ ? max_export_batch_size_ : buffer_.size();
+    }
+
+    if (num_records_to_export == 0)
+    {
+      NotifyCompletion(notify_force_flush, synchronization_data_);
+      break;
+    }
+    buffer_.Consume(num_records_to_export,
+                    [&](CircularBufferRange<AtomicUniquePtr<Recordable>> range) noexcept {
+                      range.ForEach([&](AtomicUniquePtr<Recordable> &ptr) {
+                        std::unique_ptr<Recordable> swap_ptr = std::unique_ptr<Recordable>(nullptr);
+                        ptr.Swap(swap_ptr);
+                        spans_arr.push_back(std::unique_ptr<Recordable>(swap_ptr.release()));
+                        return true;
+                      });
+                    });
+
+    exporter_->Export(nostd::span<std::unique_ptr<Recordable>>(spans_arr.data(), spans_arr.size()));
+    NotifyCompletion(notify_force_flush, synchronization_data_);
+  } while (true);
+}
+
+void BatchSpanProcessor::NotifyCompletion(
+    bool notify_force_flush,
+    const std::shared_ptr<SynchronizationData> &synchronization_data)
+{
+  if (!synchronization_data)
+  {
+    return;
+  }
+
+  if (notify_force_flush)
+  {
+    synchronization_data->is_force_flush_notified.store(true, std::memory_order_release);
+    synchronization_data->force_flush_cv.notify_one();
   }
 }
 
 void BatchSpanProcessor::DrainQueue()
 {
-  while (buffer_.empty() == false)
+  while (true)
   {
-    Export(false);
+    if (buffer_.empty() &&
+        false == synchronization_data_->is_force_flush_pending.load(std::memory_order_acquire))
+    {
+      break;
+    }
+
+    Export();
+  }
+}
+
+void BatchSpanProcessor::GetWaitAdjustedTime(
+    std::chrono::microseconds &timeout,
+    std::chrono::time_point<std::chrono::system_clock> &start_time)
+{
+  auto end_time = std::chrono::system_clock::now();
+  auto offset   = std::chrono::duration_cast<std::chrono::microseconds>(end_time - start_time);
+  start_time    = end_time;
+  timeout       = opentelemetry::common::DurationUtil::AdjustWaitForTimeout(
+      timeout, std::chrono::microseconds::zero());
+  if (timeout > offset && timeout > std::chrono::microseconds::zero())
+  {
+    timeout -= offset;
+  }
+  else
+  {
+    // Some module use zero as indefinite timeout.So we can not reset timeout to zero here
+    timeout = std::chrono::microseconds(1);
   }
 }
 
 bool BatchSpanProcessor::Shutdown(std::chrono::microseconds timeout) noexcept
 {
-  std::lock_guard<std::mutex> shutdown_guard{shutdown_m_};
-  bool already_shutdown = is_shutdown_.exchange(true);
+  auto start_time = std::chrono::system_clock::now();
+  std::lock_guard<std::mutex> shutdown_guard{synchronization_data_->shutdown_m};
+  bool already_shutdown = synchronization_data_->is_shutdown.exchange(true);
 
   if (worker_thread_.joinable())
   {
-    cv_.notify_one();
+    synchronization_data_->is_force_wakeup_background_worker.store(true, std::memory_order_release);
+    synchronization_data_->cv.notify_one();
     worker_thread_.join();
   }
 
+  GetWaitAdjustedTime(timeout, start_time);
   // Should only shutdown exporter ONCE.
   if (!already_shutdown && exporter_ != nullptr)
   {
-    return exporter_->Shutdown();
+    return exporter_->Shutdown(timeout);
   }
 
   return true;
@@ -200,7 +283,7 @@ bool BatchSpanProcessor::Shutdown(std::chrono::microseconds timeout) noexcept
 
 BatchSpanProcessor::~BatchSpanProcessor()
 {
-  if (is_shutdown_.load() == false)
+  if (synchronization_data_->is_shutdown.load() == false)
   {
     Shutdown();
   }