+++ sdk/test/logs/batch_log_processor_test.cc
@@ -9,6 +9,7 @@
 
 #  include <gtest/gtest.h>
 #  include <chrono>
+#  include <memory>
 #  include <thread>
 
 using namespace opentelemetry::sdk::logs;